requirements.txt
asgiref==3.10.0
colorama==0.4.6
coverage==7.11.3
Django==5.2.8
django-admin-interface==0.31.0
django-colorfield==0.14.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
iniconfig==2.3.0
model-bakery==1.20.5
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
